% Generated by build script
\name{treas_auctions}
\alias{treas_auctions}
\title{Treasury auction results}
\usage{
treas_auctions()
}
\arguments{
\item{security_type}{Filter: "Bill", "Note", "Bond", "TIPS", "FRN"}
\item{start}{Start date}
\item{end}{End date}
\item{max_results}{Max records (default 100)}
}
\value{tibble with auction details}
\description{Treasury auction results}
