% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treas.R
\name{treas_cash_balance}
\alias{treas_cash_balance}
\title{Daily Treasury Statement - operating cash balance}
\usage{
treas_cash_balance(start = NULL, end = NULL, max_results = 90)
}
\arguments{
\item{start}{Start date}

\item{end}{End date}

\item{max_results}{Max records (default 90)}
}
\value{
tibble: record_date, account_type, open_today_bal,
close_today_bal
}
\description{
Daily Treasury Statement - operating cash balance
}
