% Generated by build script
\name{treas_debt}
\alias{treas_debt}
\title{Total public debt outstanding (Debt to the Penny)}
\usage{
treas_debt()
}
\arguments{
\item{start}{Start date (Date or string). NULL = all.}
\item{end}{End date. NULL = today.}
\item{max_results}{Max records (default 365 = ~1 year of daily data)}
}
\value{tibble: record_date, debt_held_public_amt, intragov_hold_amt,}
\description{Total public debt outstanding (Debt to the Penny)}
