% Generated by build script
\name{treas_debt_foreign}
\alias{treas_debt_foreign}
\title{Federal debt held by foreign and international investors}
\usage{
treas_debt_foreign()
}
\arguments{
\item{max_results}{Max records (default 100)}
}
\value{tibble with country-level holdings}
\description{Federal debt held by foreign and international investors}
