% Generated by build script
\name{treas_exchange}
\alias{treas_exchange}
\title{Treasury exchange rates (quarterly)}
\usage{
treas_exchange()
}
\arguments{
\item{country}{Filter by country name (e.g. "Canada", "Japan")}
\item{start}{Start date}
\item{end}{End date}
\item{max_results}{Max records (default 500)}
}
\value{tibble: record_date, country, currency, country_currency_desc,}
\description{Treasury exchange rates (quarterly)}
