% Generated by build script
\name{treas_savings_bonds}
\alias{treas_savings_bonds}
\title{Savings bond redemption values}
\usage{
treas_savings_bonds()
}
\arguments{
\item{max_results}{Max records (default 100)}
}
\value{tibble with bond series and redemption values}
\description{Savings bond redemption values}
