% Generated by build script
\name{treas_securities}
\alias{treas_securities}
\title{Treasury securities outstanding (Monthly Statement of Public Debt)}
\usage{
treas_securities()
}
\arguments{
\item{max_results}{Max records (default 100)}
}
\value{tibble with security types and amounts outstanding}
\description{Treasury securities outstanding (Monthly Statement of Public Debt)}
