% Generated by build script
\name{treas_cash_balance}
\alias{treas_cash_balance}
\title{Daily Treasury Statement - operating cash balance}
\usage{
treas_cash_balance()
}
\arguments{
\item{start}{Start date}
\item{end}{End date}
\item{max_results}{Max records (default 90)}
}
\value{tibble: record_date, account_type, open_today_bal,}
\description{Daily Treasury Statement - operating cash balance}
