% Generated by build script
\name{treas_rates}
\alias{treas_rates}
\title{Average interest rates on Treasury securities}
\usage{
treas_rates()
}
\arguments{
\item{start}{Start date filter}
\item{end}{End date filter}
\item{security}{Optional security type filter}
\item{max_results}{Max records (default 500)}
}
\value{tibble: record_date, security_type_desc, security_desc,}
\description{Average interest rates on Treasury securities}
